Initial commit of tic-tac-toe sample
This is adapted from the ReactJS tutorial: https://reactjs.org/tutorial/tutorial.html

To ease comparison (and show exactly what had to be changed to port to Magic Script),
this commit does not alter the ReactJS code at all. This means it will not build as is.
The next commit will contain only the changes required for porting to Magic Script / Lumin.